+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - btstack_audio: added get_samplerate function to help with audio sample rate synchronization 
 - btstack_flash_bank: write empty tag instead of overwriting existing tag with ENABLE_TLV_FLASH_WRITE_ONCE
 - esp32: drop support for Makefile projects from esp-idf 3.x
+- sco_demo_util: replace hfp_msbc by hfp_codec
+
 
 ## Release v1.5.5